% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IMR90_100kb_chr20}
\alias{IMR90_100kb_chr20}
\title{Hi-C data for IMR90 cells}
\format{A 625x625 symmetric nonnegative integer-valued matrix.}
\usage{
IMR90_100kb_chr20
}
\description{
Contact matrix calculated for chromosomes 20 and resolutions 100kb (obtained from the Gene Expression Omnibus).
}
\keyword{datasets}
