% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IMR90_100kb_chr21}
\alias{IMR90_100kb_chr21}
\title{Hi-C data for IMR90 cells for chromosome 21 and resolution 100kb obtained from the Gene Expression Omnibus}
\format{A 470x470 symmetric nonnegative integer-valued matrix.}
\usage{
IMR90_100kb_chr21
}
\description{
usage data(IMR90_100kb_chr21)
}
\keyword{datasets}
