% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IMR90_50kb_chr21}
\alias{IMR90_50kb_chr21}
\title{Hi-C data for IMR90 cells for chromosomes 21 and resolutions 50kb obtained from the Gene Expression Omnibus}
\format{A 939x939 symmetric nonnegative integer-valued matrix.}
\usage{
IMR90_50kb_chr21
}
\description{
usage data(IMR90_50kb_chr21)
}
\keyword{datasets}
