% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoisMS.R
\name{PoisMS}
\alias{PoisMS}
\title{Poisson Metric Scaling}
\usage{
PoisMS(
  C,
  H,
  alpha = 1,
  beta,
  X0 = NULL,
  eps_wpcms = 1e-06,
  maxiter = 100,
  verbose_wpcms = FALSE,
  eps_poisms = 1e-06,
  maxepoch = 100,
  verbose_poisms = FALSE
)
}
\arguments{
\item{C}{a square symmetric matrix representing a Hi-C contact matrix.}

\item{H}{a spline basis matrix. By default assumed to have orthogonal columns. If not, orthogonalized via QR decomposition.}

\item{alpha, beta}{scaling and centering constants for Poisson model.}

\item{X0}{an initialization for \eqn{X} of the PoisMS algorithm. If \code{X0 = NULL}, the PCMS warm start is considered.}

\item{eps_wpcms, eps_poisms}{positive convergence tolerances for WPCMS inner loop and PoisMS outer loop.}

\item{maxiter, maxepoch}{integers giving the maximal numbers of iterations for WPCMS inner loop and PoisMS outer loop.}

\item{verbose_wpcms, verbose_poisms}{logical. If \code{TRUE}, the WPCMS loss after each iteration of inner loop and PoisMS loss after each epoch of outer loop are printed.}
}
\value{
A list containing the PoisMS problem solution:
\itemize{
  \item \code{Theta} -- the matrix of spline parameters.
  \item \code{X} -- the resulting conformation reconstruction.
  \item \code{loss} -- the resulting value of the PoisMS loss.
  \item \code{epoch} -- the total number of epochs.
  \item \code{plot} -- the convergence plot.
}
}
\description{
PoisMS function calculates the Poisson Metric Scaling solution for a contact matrix \eqn{C} and a spline basis matrix \eqn{H}.
The optimal solution is found via minimizing the negative log-likelihood for the Poisson probabilistic model \eqn{C~Pois(\Lambda)} with \deqn{\log(\Lambda) = \alpha XX^T + \beta}{log(\Lambda) = \alpha XX' + \beta} w.r.t. \eqn{\Theta} subject to the smooth curve constraint \eqn{X = H\Theta}.
The solution can be calculated via iterating the second order approximation of the objective (outer PoisMS loop) and applying WPCMS to optimize the obtained quadratic approximation (inner WPCMS loop).
The spatial coordiantes of the resulting reconstruction are presented in \eqn{X}.
}
\examples{
data(C)
data(H)
##PoisMS solution
alpha = 1
beta = log(mean(C))
PoisMS(C, H, alpha, beta)$X

}
