% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sCCA.r
\name{sCCA.binary_search}
\alias{sCCA.binary_search}
\title{Binary search for choosing the value of d}
\usage{
sCCA.binary_search(v, l, iteration = 150)
}
\arguments{
\item{v}{}

\item{l}{}
}
\value{
d
}
\description{
Binary search for choosing the value of d
}
