% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sCCA.r
\name{sCCA}
\alias{sCCA}
\title{Sparse CCA}
\usage{
sCCA(X1, X2, l1, l2, iteration = 1000, threshold = 1e-06)
}
\arguments{
\item{X1, }{X2 datasets}

\item{l1, }{l2}
}
\value{
List of weights w1, w2
}
\description{
Sparse CCA
}
\references{
Witten D. M., Tibshirani R.,  and Hastie, T. (2009)
\emph{A penalized matrix decomposition, with applications to sparse principal components and canonical correlation analysis}, \emph{Biostatistics, Gol 10 (3), 515-534, Jul 2009}\cr
}
