% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Reconstruction vizualization}
\usage{
visualize(X, index = 1:nrow(X), type = "projection")
}
\arguments{
\item{X}{a matrix representing spatial coordinates of resulting chromatin reconstruction.}

\item{index}{points where spline basis is evaluated; each corresponds to a particular genomic loci.}

\item{type}{the type of plot returned. If \code{type = 'heatmap'}, the contact matrix approximation is returned. Set \code{type = 'projection'} and \code{type = '3D'} to output the projection and 3D model of chromatin conformation reconstruction, respectively.}
}
\value{
Reconstruction plots.
}
\description{
This function allows to plot reconstructed chromatin conformation \emph{X} and corresponding contact matrix approximation \emph{XX'}.
}
\examples{
data(C)

#create spline basis matrix
H = bs(1:ncol(C), df = 5)

#orthogonalize H using QR decomposition
H = qr.Q(qr(H))

#find WPCMS reconstruction
X = WPCMS(C, H)$X

#plot contact matrix approximation
visualize(X, type = 'heatmap')

#plot projection of reconstructed chromatin conformation
visualize(X, type = 'projection')

#plot 3D model of reconstructed chromatin conformation
visualize(X, type = '3D')

}
