% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPCMS.R
\name{WPCMS}
\alias{WPCMS}
\title{Weighted Principal Curve Metric Scaling}
\usage{
WPCMS(C, H, W = NULL, X0 = NULL, eps = 1e-06, maxiter = 100, verbose = FALSE)
}
\arguments{
\item{C}{a square symmetric matrix representing a Hi-C contact matrix.}

\item{H}{a spline basis matrix. By default assumed to have orthogonal columns. If not, orthogonalized via QR decomposition.}

\item{W}{a weights matrix, same dimension as \code{C}. If \code{W = NULL}, the Principal Curve Metric Scaling solution is calculated.}

\item{X0}{an initialization for \eqn{X} of the WPCMS algorithm. If \code{X0 = NULL}, the PCMS warm start is considered.}

\item{eps}{a positive convergence tolerance.}

\item{maxiter}{an integer giving the maximal number of iterations.}

\item{verbose}{logical. If \code{TRUE}, the WPCMS loss after each iteration is printed.}
}
\value{
A list containing the WPCMS problem solution:
\itemize{
  \item \code{Theta} -- the matrix of spline parameters.
  \item \code{X} -- the resulting conformation reconstruction.
  \item \code{loss} -- the resulting value of the WPCMS loss.
  \item \code{iter} -- the total number of iterations.
}
}
\description{
WPCMS function calculates the Weighted Principal Curve Metric Scaling solution for a contact matrix \eqn{C} and a spline basis matrix \eqn{H}.
The optimal solution is found via minimizing the weighted Frobenius norm \deqn{ \|W * (C-X X^T) \|^2_F }{|| W x (C - XX') ||} w.r.t. \eqn{\Theta} subject to the smooth curve constraint \eqn{X = H\Theta}.
The spatial coordiantes of the resulting reconstruction are presented in \eqn{X}.
}
\examples{
data(C)
data(H)
##W = NULL, PCMS solution is calculated
WPCMS(C, H)$X
##WPCMS solution
W = matrix(runif(length(C)), dim(C))
WPCMS(C, H, W)$X

}
