% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rCCA}
\alias{rCCA}
\title{Computer sparse cannonical correlation vectors via Suo 2017 method}
\usage{
rCCA(X1, X2, l1, l2, niter = 1000L, threshold = 1e-06, verbose = TRUE)
}
\arguments{
\item{X1}{Matrix of covariates.}

\item{X2}{Matrix of covariates.}

\item{l1}{Penalisation term.}

\item{l2}{Penalisation term.}

\item{niter}{Number of iterations to run algorithm for (default = 1000).}

\item{threshold}{Stopping criterea threshold (default = 1e-6).}

\item{verbose}{Print debug information.}
}
\description{
Computer sparse cannonical correlation vectors via Suo 2017 method
}
