% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcca.R
\name{sCCA}
\alias{sCCA}
\alias{rCCA}
\title{Sparse Canonical Correlation Analysis}
\usage{
sCCA(
  X1,
  X2,
  l1 = 0.3 * sqrt(ncol(X1)),
  l2 = 0.3 * sqrt(ncol(X2)),
  K = 1,
  niter = 1000,
  threshold = 1e-06,
  verbose = TRUE
)

rCCA(
  X1,
  X2,
  l1 = 0.3,
  l2 = 0.3,
  K = 1,
  niter = 1000,
  threshold = 1e-06,
  verbose = TRUE
)
}
\arguments{
\item{X1, X2:}{Matrices of covariates.}

\item{l1, l2:}{Penalisation terms.}

\item{niter}{Number of iterations to run algorithm for (default = 1000).}

\item{threshold:}{Stopping criterea threshold (default = 1e-6).}

\item{verbose:}{Print debug information.}
}
\description{
Sparse Canonical Correlation Analysis
}
